+++ Web project/src/beans/Restaurant.java
@@ -15,14 +15,14 @@ public class Restaurant {
 	private Address address;
 	private String link;
 	private Boolean deleted;
-	private int grade;
+	private double grade;
 	
 	
-	public int getGrade() {
+	public double getGrade() {
 		return grade;
 	}
 
-	public void setGrade(int grade) {
+	public void setGrade(double grade) {
 		this.grade = grade;
 	}
 
@@ -88,7 +88,7 @@ public class Restaurant {
 	}
 
 	public Restaurant(int id,String managerId, String name, RestaurantType type, ArrayList<Integer> articalIds, Status status,
-			Address address, String link, Boolean deleted, int grade) {
+			Address address, String link, Boolean deleted, double grade) {
 		super();
 		this.id = id;
 		this.managerId = managerId;
